requirements.txt

lxml
pandas
sqlalchemy>=2.0
psycopg[binary,pool]
python-dotenv
tqdm
pyyaml
scikit-learn
lightgbm
xgboost
matplotlib
seaborn
requests
pre-commit
ruff
black
pytest
